add type for geocoder and update imports at all places
